+++ replayer/src/schema.rs
@@ -2,7 +2,6 @@ use serde_derive::{Deserialize, Serialize};
 use serde_json::Value;
 use replay_engine::account_data_store::AccountDataStore;
 use replay_engine::decoded_instructions::{deserialize_u64, deserialize_base64, serialize_base64};
-use crate::util::{serialize_account_data_store, deserialize_account_data_store_on_disk, deserialize_account_data_store_on_memory};
 
 pub use replay_engine::decoded_instructions::{DecodedInstruction, DecodedProgramDeployInstruction, DecodedWhirlpoolInstruction};
 
@@ -32,42 +31,6 @@ pub struct WhirlpoolState {
   pub block_time: i64,
   pub accounts: AccountDataStore,
   pub program_data: Vec<u8>,
-}
-
-#[derive(Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct WhirlpoolStateOnMemoryDeserializer {
-  pub slot: u64,
-  pub block_height: u64,
-  pub block_time: i64,
-  #[serde(deserialize_with = "deserialize_account_data_store_on_memory")]
-  pub accounts: AccountDataStore,
-  #[serde(deserialize_with = "deserialize_base64")]
-  pub program_data: Vec<u8>,
-}
-
-#[derive(Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct WhirlpoolStateOnDiskDeserializer {
-  pub slot: u64,
-  pub block_height: u64,
-  pub block_time: i64,
-  #[serde(deserialize_with = "deserialize_account_data_store_on_disk")]
-  pub accounts: AccountDataStore,
-  #[serde(deserialize_with = "deserialize_base64")]
-  pub program_data: Vec<u8>,
-}
-
-#[derive(Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct WhirlpoolStateSerializer<'a> {
-  pub slot: u64,
-  pub block_height: u64,
-  pub block_time: i64,
-  #[serde(serialize_with = "serialize_account_data_store")]
-  pub accounts: &'a AccountDataStore,
-  #[serde(serialize_with = "serialize_base64")]
-  pub program_data: &'a Vec<u8>,
 }
 
 #[derive(Serialize, Deserialize, Debug, PartialEq, Eq)]
